+++ app/components/product/index.ts
@@ -8,5 +8,4 @@ export { default as StickyPurchaseFooterMobile } from "./StickyPurchaseFooterMob
 export { default as ClinicalIngredients } from "./ClinicalIngredients";
 export { default as FormulaBenefitsPillars } from "./FormulaBenefitsPillars";
 export { default as ProductWhatYouGet } from "./ProductWhatYouGet";
-export { default as FormulaFAQ } from "./FormulaFAQ";